+++ NEWS.md
@@ -1,3 +1,4 @@
+-   climfor 0.1.0.9004: chirps & gshtd zarr dl & dask+slurm
 -   climfor 0.1.0.9003: tmf deforest & kendall's tau for chirps
 -   climfor 0.1.0.9002: quick exploration with dl script def
 -   climfor 0.1.0.9001: init